# docs/swagger.yaml
definitions:
  models.ApiResponse-array_models_WorkTeam:
    properties:
      data:
        description: Response data
        items:
          $ref: '#/definitions/models.WorkTeam'
        type: array
      message:
        description: Descriptive message
        example: Successful operation
        type: string
      status:
        description: Response status
        example: success
        type: string
    type: object
  models.ApiResponse-models_WorkTeam:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/models.WorkTeam'
        description: Response data
      message:
        description: Descriptive message
        example: Successful operation
        type: string
      status:
        description: Response status
        example: success
        type: string
    type: object
  models.WorkTeam:
    description: Structure representing a work team
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      description:
        example: Team responsible for software development
        type: string
      members:
        example:
        - '["user1"'
        - ' "user2"]'
        items:
          type: string
        type: array
      status:
        example: true
        type: boolean
      team_code:
        example: TEAM-001
        type: string
      team_name:
        example: Development Team
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      work_team_id:
        example: 1
        type: integer
    type: object
host: localhost:8180
info:
  contact: {}
  description: API for managing work teams.
  title: Work Team Search API
  version: "1.0"
paths:
  /work-team:
    get:
      description: Get all the work team
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiResponse-array_models_WorkTeam'
      summary: List teams
      tags:
      - work-team
  /work-team/{team_code}:
    get:
      consumes:
      - application/json
      description: Get a specific work team by its unique code
      parameters:
      - description: Unique team code
        in: path
        name: team_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiResponse-models_WorkTeam'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiResponse-models_WorkTeam'
      summary: Get work team by code
      tags:
      - work-team
schemes:
- http
- https
swagger: "2.0"